DT-158 Architecture detected

+++ lib/core/rancher/step.sh
@@ -50,7 +50,9 @@ function ih::setup::core.rancher::deps() {
 function ih::setup::core.rancher::install() {
     local THIS_DIR="$IH_CORE_LIB_DIR/core/rancher"
 
-    cp "${THIS_DIR}/io.rancherdesktop.profile.defaults.plist" "$HOME/Library/Preferences/io.rancherdesktop.profile.defaults.plist"
+    echo "A configuration file for Rancher Desktop will be copied to your system"
+    echo "You may be required to enter your password"
+    sudo cp "${THIS_DIR}/io.rancherdesktop.profile.defaults.plist" "$HOME/Library/Preferences/io.rancherdesktop.profile.defaults.plist"
 
     # Check if Rancher was installed manually
     brew list rancher > /dev/null 2>&1
@@ -80,10 +82,9 @@ function ih::setup::core.rancher::install() {
     # Installation and configuration of Rancher Desktop
     for _ in 1 2 3; do
 
-        # Detect Rosetta
-        if [[ $(sysctl -n sysctl.proc_translated) -eq 1 ]]; then
-            # Rosetta  Active
-            arch -arm64 -c brew reinstall ih-rancher
+        # Check if we have a Mac M1 and the terminal is running  over x86
+        if [[ $(sysctl -n sysctl.proc_translated) -eq 1 ]] && [ $(arch) = "i386" ]; then
+            mbrew reinstall ih-rancher
         else
             brew reinstall ih-rancher
         fi
@@ -94,21 +95,27 @@ function ih::setup::core.rancher::install() {
         fi
     done
 
-    rm -rf ~/.rd
-    $(/Applications/Rancher\ Desktop.app/Contents/Resources/resources/darwin/bin/rdctl start)
-    # Check for docker binary in /usr/local/bin
-    # Check if /usr/local/bin/docker exists
-    DOCKERBIN=/usr/local/bin/docker
-    DOCKERCOMPOSEBIN=/usr/local/bin/docker-compose
-    if [ ! -f "$DOCKERBIN" ]; then
+    # Continue with the setting just if the cask was installed successfully
+    if [ $CASKSUCCEEDED -eq 0 ]; then
+        rm -rf ~/.rd
+        $(/Applications/Rancher\ Desktop.app/Contents/Resources/resources/darwin/bin/rdctl start)
+        # Check for docker binary in /usr/local/bin
+        # Check if /usr/local/bin/docker exists
+        DOCKERBIN=/usr/local/bin/docker
+        DOCKERCOMPOSEBIN=/usr/local/bin/docker-compose
+        if [ ! -f "$DOCKERBIN" ]; then
 
-        echo "In order to continue with Rancher configuration and be able to use this engine, some IDEs require the creation of symlinks for remote Python interpreters"
-        echo "Your password is required for the creation of symlink mentioned above"
-        sudo ln -s $HOME/.rd/bin/docker /usr/local/bin/docker
-        if [ ! -f "$DOCKERCOMPOSEBIN" ]; then
-            sudo ln -s $HOME/.rd/bin/docker-compose /usr/local/bin/docker-compose
+            echo "In order to continue with Rancher configuration and be able to use this engine, some IDEs require the creation of symlinks for remote Python interpreters"
+            echo "Your password is required for the creation of symlink mentioned above"
+            sudo ln -s $HOME/.rd/bin/docker /usr/local/bin/docker
+            if [ ! -f "$DOCKERCOMPOSEBIN" ]; then
+                sudo ln -s $HOME/.rd/bin/docker-compose /usr/local/bin/docker-compose
+            fi
         fi
+        
+        echo "Rancher Desktop has been installed successfully"
+    else
+        echo "There was an error with Rancher Desktop Installation. Please contact  support
+            in the #developer-tools channel in Slack (https://ih-epdd.slack.com/archives/C04LPMF4YPL)" 
     fi
-    
-    echo "Rancher Desktop has been installed successfully"
 }
